网络：迁移vcpkg依赖路径并修复curl错误格式化

+++ kbe/src/lib/network/http_utility.cpp
@@ -47,7 +47,7 @@ bool initialize()
 		if (CURLE_OK != curlCode)
 		{
 			ERROR_MSG(fmt::format("Http::initialize: "
-				"curl_global_init error! curlCode={}\n", curlCode));
+				"curl_global_init error! curlCode={}\n", static_cast<int>(curlCode)));
 
 			return false;
 		}
@@ -133,7 +133,7 @@ Request::Status Request::setCAInfo(const std::string& cainfo)
 	CURLcode curlCode = curl_easy_setopt((CURL*)pContext_, CURLOPT_CAINFO, cainfo.c_str());
 	if (CURLE_OK != curlCode)
 	{
-		ERROR_MSG(fmt::format("Http::Request::setCAInfo: error! curlCode={}\n", curlCode));
+		ERROR_MSG(fmt::format("Http::Request::setCAInfo: error! curlCode={}\n", static_cast<int>(curlCode)));
 		return INVALID_OPT;
 	}
 
@@ -146,7 +146,7 @@ Request::Status Request::setSSLCert(const std::string& sslCert)
 	CURLcode curlCode = curl_easy_setopt((CURL*)pContext_, CURLOPT_SSLCERT, sslCert.c_str());
 	if (CURLE_OK != curlCode)
 	{
-		ERROR_MSG(fmt::format("Http::Request::setSSLCert: error! curlCode={}\n", curlCode));
+		ERROR_MSG(fmt::format("Http::Request::setSSLCert: error! curlCode={}\n", static_cast<int>(curlCode)));
 		return INVALID_OPT;
 	}
 
@@ -159,7 +159,7 @@ Request::Status Request::setSSLKey(const std::string& sslKey)
 	CURLcode curlCode = curl_easy_setopt((CURL*)pContext_, CURLOPT_SSLKEY, sslKey.c_str());
 	if (CURLE_OK != curlCode)
 	{
-		ERROR_MSG(fmt::format("Http::Request::setSSLKey: error! curlCode={}\n", curlCode));
+		ERROR_MSG(fmt::format("Http::Request::setSSLKey: error! curlCode={}\n", static_cast<int>(curlCode)));
 		return INVALID_OPT;
 	}
 
@@ -172,7 +172,7 @@ Request::Status Request::setSSLKeyPassword(const std::string& sslKeyPwd)
 	CURLcode curlCode = curl_easy_setopt((CURL*)pContext_, CURLOPT_KEYPASSWD, sslKeyPwd.c_str());
 	if (CURLE_OK != curlCode)
 	{
-		ERROR_MSG(fmt::format("Http::Request::setSSLKeyPassword: error! curlCode={}\n", curlCode));
+		ERROR_MSG(fmt::format("Http::Request::setSSLKeyPassword: error! curlCode={}\n", static_cast<int>(curlCode)));
 		return INVALID_OPT;
 	}
 
@@ -187,7 +187,7 @@ Request::Status Request::setSSLVerifyPeer(long v)
 
 	if (CURLE_OK != curlCode)
 	{
-		ERROR_MSG(fmt::format("Http::Request::setSSLVerifyPeer: error! curlCode={}\n", curlCode));
+		ERROR_MSG(fmt::format("Http::Request::setSSLVerifyPeer: error! curlCode={}\n", static_cast<int>(curlCode)));
 		return INVALID_OPT;
 	}
 
@@ -202,7 +202,7 @@ Request::Status Request::setSSLVerifyHost(long v)
 
 	if (CURLE_OK != curlCode)
 	{
-		ERROR_MSG(fmt::format("Http::Request::setSSLVerifyHost: error! curlCode={}\n", curlCode));
+		ERROR_MSG(fmt::format("Http::Request::setSSLVerifyHost: error! curlCode={}\n", static_cast<int>(curlCode)));
 		return INVALID_OPT;
 	}
 
@@ -224,7 +224,7 @@ Request::Status Request::setURL(const std::string& url)
 	if (CURLE_OK != curlCode)
 	{
 		ERROR_MSG(fmt::format("Http::Request::setURL: "
-			"curl_easy_setopt(CURLOPT_URL) error! curlCode={}, url={}\n", curlCode, url));
+			"curl_easy_setopt(CURLOPT_URL) error! curlCode={}, url={}\n", static_cast<int>(curlCode), url));
 
 		return INVALID_OPT;
 	}
@@ -248,7 +248,7 @@ Request::Status Request::setFollowURL(int maxRedirs)
 		if (CURLE_OK != curlCode)
 		{
 			ERROR_MSG(fmt::format("Http::Request::setFollowURL: "
-				"curl_easy_setopt(CURLOPT_MAXREDIRS) error! curlCode={}, maxRedirs={}\n", curlCode, maxRedirs));
+				"curl_easy_setopt(CURLOPT_MAXREDIRS) error! curlCode={}, maxRedirs={}\n", static_cast<int>(curlCode), maxRedirs));
 
 			return INVALID_OPT;
 		}
@@ -259,7 +259,7 @@ Request::Status Request::setFollowURL(int maxRedirs)
 	if (CURLE_OK != curlCode)
 	{
 		ERROR_MSG(fmt::format("Http::Request::setFollowURL: "
-			"curl_easy_setopt(CURLOPT_FOLLOWLOCATION) error! curlCode={}, maxRedirs={}\n", curlCode, maxRedirs));
+			"curl_easy_setopt(CURLOPT_FOLLOWLOCATION) error! curlCode={}, maxRedirs={}\n", static_cast<int>(curlCode), maxRedirs));
 
 		return INVALID_OPT;
 	}
@@ -285,7 +285,7 @@ Request::Status Request::setPostData(const void* data, unsigned int size)
 	if (CURLE_OK != curlCode)
 	{
 		ERROR_MSG(fmt::format("Http::Request::setPostData: "
-			"curl_easy_setopt(CURLOPT_POST) error! curlCode={}, size={}\n", curlCode, size));
+			"curl_easy_setopt(CURLOPT_POST) error! curlCode={}, size={}\n", static_cast<int>(curlCode), size));
 
 		return INVALID_OPT;
 	}
@@ -307,7 +307,7 @@ Request::Status Request::setPostData(const void* data, unsigned int size)
 	if (CURLE_OK != curlCode)
 	{
 		ERROR_MSG(fmt::format("Http::Request::setPostData: "
-			"curl_easy_setopt(CURLOPT_POSTFIELDS) error! curlCode={}, size={}\n", curlCode, size));
+			"curl_easy_setopt(CURLOPT_POSTFIELDS) error! curlCode={}, size={}\n", static_cast<int>(curlCode), size));
 
 		return INVALID_OPT;
 	}
@@ -316,7 +316,7 @@ Request::Status Request::setPostData(const void* data, unsigned int size)
 	if (CURLE_OK != curlCode)
 	{
 		ERROR_MSG(fmt::format("Http::Request::setPostData: "
-			"curl_easy_setopt(CURLOPT_POSTFIELDSIZE) error! curlCode={}, size={}\n", curlCode, size));
+			"curl_easy_setopt(CURLOPT_POSTFIELDSIZE) error! curlCode={}, size={}\n", static_cast<int>(curlCode), size));
 
 		return INVALID_OPT;
 	}
@@ -345,7 +345,7 @@ Request::Status Request::setHeader(const std::string& header)
 	if (CURLE_OK != curlCode)
 	{
 		ERROR_MSG(fmt::format("Http::Request::setHeader: "
-			"curl_easy_setopt(CURLOPT_HTTPHEADER) error! curlCode={}, header={}\n", curlCode, header));
+			"curl_easy_setopt(CURLOPT_HTTPHEADER) error! curlCode={}, header={}\n", static_cast<int>(curlCode), header));
 
 		return INVALID_OPT;
 	}
@@ -375,7 +375,7 @@ Request::Status Request::setHeader(const std::vector<std::string>& headers)
 		if (!headers_)
 		{
 			ERROR_MSG(fmt::format("Http::Request::setHeaders: "
-				"curl_slist_append error! curlCode={}, header={}\n", curlCode, (*iter)));
+				"curl_slist_append error! curlCode={}, header={}\n", static_cast<int>(curlCode), (*iter)));
 
 			return INVALID_OPT;
 		}
@@ -386,7 +386,7 @@ Request::Status Request::setHeader(const std::vector<std::string>& headers)
 	if (CURLE_OK != curlCode)
 	{
 		ERROR_MSG(fmt::format("Http::Request::setHeaders: "
-			"curl_easy_setopt(CURLOPT_HTTPHEADER) error! curlCode={}\n", curlCode));
+			"curl_easy_setopt(CURLOPT_HTTPHEADER) error! curlCode={}\n", static_cast<int>(curlCode)));
 
 		return INVALID_OPT;
 	}
@@ -420,7 +420,7 @@ Request::Status Request::setHeader(const std::map<std::string, std::string>& hea
 		if (!headers_)
 		{
 			ERROR_MSG(fmt::format("Http::Request::setHeaders: "
-				"curl_slist_append error! curlCode={}, header={}\n", curlCode, header));
+				"curl_slist_append error! curlCode={}, header={}\n", static_cast<int>(curlCode), header));
 
 			return INVALID_OPT;
 		}
@@ -431,7 +431,7 @@ Request::Status Request::setHeader(const std::map<std::string, std::string>& hea
 	if (CURLE_OK != curlCode)
 	{
 		ERROR_MSG(fmt::format("Http::Request::setHeaders: "
-			"curl_easy_setopt(CURLOPT_HTTPHEADER) error! curlCode={}\n", curlCode));
+			"curl_easy_setopt(CURLOPT_HTTPHEADER) error! curlCode={}\n", static_cast<int>(curlCode)));
 
 		return INVALID_OPT;
 	}
@@ -446,7 +446,7 @@ Request::Status Request::setTimeout(uint32 secs)
 	if (CURLE_OK != curlCode)
 	{
 		ERROR_MSG(fmt::format("Http::Request::setTimeout: "
-			"curl_easy_setopt(CURLOPT_TIMEOUT) error! curlCode={}, secs={}\n", curlCode, secs));
+			"curl_easy_setopt(CURLOPT_TIMEOUT) error! curlCode={}, secs={}\n", static_cast<int>(curlCode), secs));
 
 		return INVALID_OPT;
 	}
@@ -461,7 +461,7 @@ Request::Status Request::setProxy(const std::string& proxyIP, long proxyPort)
 	if (CURLE_OK != curlCode)
 	{
 		ERROR_MSG(fmt::format("Http::Request::setProxy: "
-			"curl_easy_setopt(CURLOPT_PROXYPORT) error! curlCode={}\n", curlCode));
+			"curl_easy_setopt(CURLOPT_PROXYPORT) error! curlCode={}\n", static_cast<int>(curlCode)));
 
 		return INVALID_OPT;
 	}
@@ -470,7 +470,7 @@ Request::Status Request::setProxy(const std::string& proxyIP, long proxyPort)
 	if (CURLE_OK != curlCode)
 	{
 		ERROR_MSG(fmt::format("Http::Request::setProxy: "
-			"curl_easy_setopt(CURLOPT_PROXYPORT) error! curlCode={}\n", curlCode));
+			"curl_easy_setopt(CURLOPT_PROXYPORT) error! curlCode={}\n", static_cast<int>(curlCode)));
 
 		return INVALID_OPT;
 	}
@@ -543,7 +543,7 @@ bool Request::updateHttpCode()
 	if (CURLE_OK != curlCode)
 	{
 		ERROR_MSG(fmt::format("Http::Request::perform: "
-			"curl_easy_getinfo(CURLINFO_RESPONSE_CODE) error! curlCode={}\n", curlCode));
+			"curl_easy_getinfo(CURLINFO_RESPONSE_CODE) error! curlCode={}\n", static_cast<int>(curlCode)));
 
 		return false;
 	}
